requirements.txt
streamlit==1.33.0
pandas==2.2.2
numpy==1.26.4
pyyaml==6.0.1
requests==2.31.0
llama-index-core==0.10.30
llama-index-embeddings-gemini==0.1.6
llama-index-embeddings-openai==0.1.8
llama-index-legacy==0.9.48
llama-index-llms-gemini==0.1.7
llama-index-llms-openai==0.1.16
llamaindex-py-client==0.1.18
typing-extensions==4.0.0  # If you are using Python version less than 3.8
datetime==4.4  # Ensure this version is valid or use an alternative if needed
